Drop GlobalQueryHandler.applyUsageScore
This method has been broken since its introduction. Nobody complained so
probably nobody used it. As such remove it without API version increase.
Otherwise all plugins would have to change their version.

The problem with this method is that in c++ space it takes a reference to the
vector of rank items and adjusts the scores in-place. PyBind however casts the
passed python list to a vector and as such the changes applied to the vector do
not propagate to the py::list on the python side.